# Conflicts:
#	megatron/core/distributed/param_and_grad_buffer.py
#	megatron/core/extensions/transformer_engine.py
#	megatron/core/pipeline_parallel/schedules.py
#	megatron/inference/text_generation/mcore_engine_server.py

+++ tests/functional_tests/shell_test_utils/run_ci_test.sh
@@ -2,14 +2,28 @@
 
 set -exo pipefail
 
-echo "------ARGUMENTS LIST --------"
-for ARGUMENT in "$@"; do
-    echo $ARGUMENT
-    KEY=$(echo $ARGUMENT | cut -f1 -d=)
+# Increase soft limit for number of open files to match hard limit
+ulimit -Sn $(ulimit -Hn)
 
-    KEY_LENGTH=${#KEY}
-    VALUE=$(eval echo ${ARGUMENT:$KEY_LENGTH+1})
-    export "$KEY"="$VALUE"
+# Increase soft limit for number of processes to match hard limit
+ulimit -Su $(ulimit -Hu)
+
+echo "------ARGUMENTS LIST --------"
+# Use eval to properly handle quoted arguments
+eval "set -- $@"
+for ARGUMENT in "$@"; do
+    # Split on first = only, preserving any subsequent = signs in the value
+    KEY="${ARGUMENT%%=*}"
+    VALUE="${ARGUMENT#*=}"
+
+    # Remove any surrounding quotes from the value if they exist
+    VALUE="${VALUE%\"}"
+    VALUE="${VALUE#\"}"
+    VALUE="${VALUE%\'}"
+    VALUE="${VALUE#\'}"
+
+    # Properly quote the value to preserve spaces and special characters
+    export "$KEY"="$(eval echo $VALUE)"
     echo "$KEY=$VALUE"
 done
 echo "---------------------------------"
@@ -25,6 +39,7 @@ MANDATORY_VARS=(
     "CHECKPOINT_LOAD_PATH"
     "DATA_PATH"
     "DATA_CACHE_PATH"
+    "ENABLE_LIGHTWEIGHT_MODE"
 )
 for mandatory_var in "${MANDATORY_VARS[@]}"; do
     if [[ -z "${!mandatory_var}" ]]; then
@@ -36,6 +51,10 @@ done
 # Extract settings from params file
 TEST_TYPE=$(cat $TRAINING_PARAMS_PATH |
     yq '.TEST_TYPE')
+MODE=$(cat $TRAINING_PARAMS_PATH |
+    yq '.MODE // "pretraining"')
+
+MODES=("pretraining" "inference")
 TEST_TYPES=("regular" "ckpt-resume" "frozen-resume" "frozen-start" "release")
 
 if [[ "$TEST_TYPE" == "release" ]]; then
@@ -56,33 +75,39 @@ IS_NEMO_TEST=$([[ $(echo "$TRAINING_SCRIPT_PATH" | tr '[:upper:]' '[:lower:]') =
 export IS_NEMO_TEST
 
 # Adjust model_config for lightweight mode
-if [[ "$ENABLE_LIGHTWEIGHT_MODE" == "true" && "$IS_NEMO_TEST" == "true" ]]; then
-    yq -i '.MODEL_ARGS."trainer.max_steps" = 2' $TRAINING_PARAMS_PATH
-    TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
-        yq '.MODEL_ARGS."trainer.max_steps // "100"')
+if [[ "$MODE" == "pretraining" && "$TEST_TYPE" != "release" ]]; then
+    if [[ "$ENABLE_LIGHTWEIGHT_MODE" == "true" && "$IS_NEMO_TEST" == "true" ]]; then
+        yq -i '.MODEL_ARGS."trainer.max_steps" = 2' $TRAINING_PARAMS_PATH
+        TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
+            yq '.MODEL_ARGS."trainer.max_steps // "100"')
 
-    N_REPEAT=1
+        N_REPEAT=1
 
-elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "true" && "$IS_NEMO_TEST" == "false" ]]; then
-    yq -i '.ENV_VARS."SKIP_PYTEST" = 1' $TRAINING_PARAMS_PATH
-    yq -i '.MODEL_ARGS."--exit-interval" = 4' $TRAINING_PARAMS_PATH
-    TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
-        yq '.MODEL_ARGS."--exit-interval" // "100"')
-    N_REPEAT=1
+    elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "true" && "$IS_NEMO_TEST" == "false" ]]; then
+        yq -i '.ENV_VARS."SKIP_PYTEST" = 1' $TRAINING_PARAMS_PATH
+        yq -i '.MODEL_ARGS."--exit-interval" = 4' $TRAINING_PARAMS_PATH
+        TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
+            yq '.MODEL_ARGS."--exit-interval" // "100"')
+        N_REPEAT=1
 
-    if [[ "$TEST_TYPE" == "ckpt-resume" || "$TEST_TYPE" == "frozen-resume" ]]; then
-        yq -i '.MODEL_ARGS."--save-interval" = 2' $TRAINING_PARAMS_PATH
+        if [[ "$TEST_TYPE" == "ckpt-resume" || "$TEST_TYPE" == "frozen-resume" ]]; then
+            yq -i '.MODEL_ARGS."--save-interval" = 2' $TRAINING_PARAMS_PATH
+        fi
+
+    elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "false" && "$IS_NEMO_TEST" == "true" ]]; then
+        TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
+            yq '.MODEL_ARGS."trainer.max_steps" // "100"')
+
+    elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "false" && "$IS_NEMO_TEST" == "false" ]]; then
+        yq -i '.MODEL_ARGS."--exit-interval" = .MODEL_ARGS."--train-iters"' $TRAINING_PARAMS_PATH
+        TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
+            yq '.MODEL_ARGS."--exit-interval" // "100"')
     fi
+fi
 
-elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "false" && "$IS_NEMO_TEST" == "true" ]]; then
-    TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
-        yq '.MODEL_ARGS."trainer.max_steps" // "100"')
-
-elif [[ "$ENABLE_LIGHTWEIGHT_MODE" == "false" && "$IS_NEMO_TEST" == "false" ]]; then
-    yq -i '.MODEL_ARGS."--exit-interval" = .MODEL_ARGS."--train-iters"' $TRAINING_PARAMS_PATH
+if [[ "$MODE" == "pretraining" && "$TEST_TYPE" = "release" ]]; then
     TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
         yq '.MODEL_ARGS."--exit-interval" // "100"')
-
 fi
 
 # Extract settings from params file
@@ -104,6 +129,7 @@ for i in $(seq 1 $N_REPEAT); do
     export RUN_NUMBER=1
     export REPEAT=$i
     export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
+    export TRAINING_EXIT_CODE=0
 
     if [[ "$TEST_TYPE" = "frozen-start" ]]; then
         export CHECKPOINT_LOAD_PATH=$_CHECKPOINT_LOAD_PATH
@@ -116,7 +142,7 @@ for i in $(seq 1 $N_REPEAT); do
         export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
     fi
 
-    bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+    bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
 
     if [[ "$TEST_TYPE" = "frozen-resume" && -z "$(ls -A "$_CHECKPOINT_LOAD_PATH" 2>/dev/null)" ]]; then
         echo "No frozen checkpoint found. Will skip second run."
@@ -127,24 +153,24 @@ for i in $(seq 1 $N_REPEAT); do
         break
     fi
 
-    if [[ "$TEST_TYPE" == "ckpt-resume" ]]; then
+    if [[ "$TEST_TYPE" == "ckpt-resume" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
         export CHECKPOINT_LOAD_PATH=$CHECKPOINT_SAVE_PATH
 
         rm -rf "$CHECKPOINT_LOAD_PATH/iter_$(printf "%07d\n" "$TRAIN_ITERS")"
         echo $((TRAIN_ITERS / 2)) >$CHECKPOINT_LOAD_PATH/latest_checkpointed_iteration.txt
 
         export RUN_NUMBER=2
-        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
     fi
 
-    if [[ "$TEST_TYPE" == "frozen-resume" ]]; then
+    if [[ "$TEST_TYPE" == "frozen-resume" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
 
         # Checkpoint-resume tests load from prev run
         export CHECKPOINT_LOAD_PATH=$_CHECKPOINT_LOAD_PATH
         export CHECKPOINT_SAVE_PATH=/tmp/checkpoints/
 
         export RUN_NUMBER=2
-        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
 
         export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
         rm -rf "$CHECKPOINT_SAVE_PATH/iter_0000$TRAIN_ITERS"
@@ -153,7 +179,6 @@ for i in $(seq 1 $N_REPEAT); do
 
     if [[ "$TEST_TYPE" == "release" ]]; then
         SKIP_PYTEST=1
-        TRAIN_ITERS=10000000
     fi
 
     if [[ ${RECORD_CHECKPOINTS} == "true" ]]; then
@@ -170,40 +195,72 @@ for i in $(seq 1 $N_REPEAT); do
             EXTRACT_ARGS=("--is-normal-test")
         fi
 
-        python3 $ROOT_DIR/tests/functional_tests/python_test_utils/get_test_results_from_tensorboard_logs.py \
-            --logs-dir $TENSORBOARD_PATH \
-            --train-iters $TRAIN_ITERS \
-            --output-path ${OUTPUT_PATH}/$(basename $GOLDEN_VALUES_PATH) \
-            "${EXTRACT_ARGS[@]}"
+        # Read test values from Tensorboard for non-inference tests.
+        # Inference tests will load from JSON instead.
+        if [[ "$MODE" == "pretraining" ]]; then
+            uv run python $ROOT_DIR/tests/functional_tests/python_test_utils/get_test_results_from_tensorboard_logs.py \
+                --logs-dir $TENSORBOARD_PATH \
+                --train-iters $TRAIN_ITERS \
+                --output-path ${OUTPUT_PATH}/$(basename $GOLDEN_VALUES_PATH) \
+                "${EXTRACT_ARGS[@]}"
+        fi
     fi
 
     # Maybe run tests
-    if [[ ${SKIP_PYTEST:-0} != 1 ]]; then
+    if [[ ${SKIP_PYTEST:-0} == 1 ]]; then
+        echo Skipping Pytest checks.
+        exit ${TRAINING_EXIT_CODE}
+    fi
 
-        export NVTE_ALLOW_NONDETERMINISTIC_ALGO
-        if [[ "${NVTE_ALLOW_NONDETERMINISTIC_ALGO}" == "1" ]]; then
-            ALLOW_NONDETERMINISTIC_ALGO_ARG="--allow-nondeterministic-algo"
-        fi
+    if [[ ! " ${TEST_TYPES[*]} " =~ " ${TEST_TYPE} " ]]; then
+        echo "Test type $TEST_TYPE not yet implemented."
+    fi
 
-        echo "Running pytest checks against golden values"
+    if [[ ! " ${MODES[*]} " =~ " ${MODE} " ]]; then
+        echo "Mode $MODE not yet implemented."
+    fi
 
-        pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_regular_pipeline.py \
+    export NVTE_ALLOW_NONDETERMINISTIC_ALGO
+    if [[ "${NVTE_ALLOW_NONDETERMINISTIC_ALGO}" == "1" ]]; then
+        ALLOW_NONDETERMINISTIC_ALGO_ARG="--allow-nondeterministic-algo"
+    fi
+
+    echo "Running pytest checks against golden values"
+
+    # For pretraining jobs
+    if [[ "$MODE" == "pretraining" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
+        uv run pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_pretraining_regular_pipeline.py \
             --golden-values-path $GOLDEN_VALUES_PATH \
             --tensorboard-path $TENSORBOARD_PATH \
+            --train-iters $TRAIN_ITERS \
             --model-config-path ${TRAINING_PARAMS_PATH} \
             $ALLOW_NONDETERMINISTIC_ALGO_ARG
 
         if [[ "$TEST_TYPE" == "ckpt-resume" || "$TEST_TYPE" == "frozen-resume" ]]; then
             echo "Running pytest 1st vs 2nd run comparison"
-            pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_resume_checkpoint_pipeline.py \
+            uv run pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_pretraining_resume_checkpoint_pipeline.py \
                 --tensorboard-path $TENSORBOARD_PATH \
                 --train-iters $TRAIN_ITERS \
                 --model-config-path ${TRAINING_PARAMS_PATH} \
                 $ALLOW_NONDETERMINISTIC_ALGO_ARG
         fi
+    fi
 
-        if [[ ! " ${TEST_TYPES[*]} " =~ " ${TEST_TYPE} " ]]; then
-            echo "Test type $TEST_TYPE not yet implemented."
+    # For inference jobs
+    if [[ "$MODE" == "inference" ]]; then
+        if [[ "$TEST_TYPE" == "frozen-start" ]]; then
+            uv run pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_inference_regular_pipeline.py \
+                --golden-values-path $GOLDEN_VALUES_PATH \
+                --test-values-path $TENSORBOARD_PATH \
+                --model-config-path ${TRAINING_PARAMS_PATH} \
+                $ALLOW_NONDETERMINISTIC_ALGO_ARG
         fi
     fi
+
+    # Abort if training failed
+    if [[ "$TRAINING_EXIT_CODE" -ne 0 && "$TEST_TYPE" != "release" ]]; then
+        echo "Training failed. Aborting."
+        exit 1
+    fi
+
 done
